add instruction to add Anaconda to PATH on Windows
This is necessary as we are also teaching Bash. Git Bash allows us
to teach Bash, but does not allow us to use Anaconda unless Anaconda
is in the PATH.

Anaconda Prompt allows us to use Anaconda, but does not allow us
to use Bash.

The compromise is to use Git Bash for the terminal.